+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog LaravelVariables
 
+## 1.2.0 - 2018-05-18
+
+- Add json decode method && helper
+- Add to config var cache name
+
 ## 1.0.1 - 2017-12-21
 
 - Update required